[git] Split out pkg utility functions to their own file
