+++ git/util.go
@@ -1,0 +1,64 @@
+package git
+
+import (
+	"bytes"
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/G-Node/gin-cli/util"
+	humanize "github.com/dustin/go-humanize"
+)
+
+// General utility functions for the git and git-annex shell commands and their output.
+
+func makeFileList(header string, fnames []string) string {
+	if len(fnames) == 0 {
+		return ""
+	}
+	var filelist bytes.Buffer
+	_, _ = filelist.WriteString(fmt.Sprintf("%s (%d)\n", header, len(fnames)))
+	for idx, name := range fnames {
+		_, _ = filelist.WriteString(fmt.Sprintf("  %d: %s\n", idx+1, name))
+	}
+	_, _ = filelist.WriteString("\n")
+	return filelist.String()
+}
+
+func calcRate(dbytes int, dt time.Duration) string {
+	dtns := dt.Nanoseconds()
+	if dtns <= 0 || dbytes <= 0 {
+		return ""
+	}
+	rate := int64(dbytes) * 1000000000 / dtns
+	return fmt.Sprintf("%s/s", humanize.IBytes(uint64(rate)))
+}
+
+func logstd(out, err []byte) {
+	util.LogWrite("[stdout]\n%s\n[stderr]\n%s", string(out), string(err))
+}
+
+func splitRepoParts(repoPath string) (repoOwner, repoName string) {
+	repoPathParts := strings.SplitN(repoPath, "/", 2)
+	repoOwner = repoPathParts[0]
+	repoName = repoPathParts[1]
+	return
+}
+
+func cutline(b []byte) (string, bool) {
+	idx := -1
+	cridx := bytes.IndexByte(b, '\r')
+	nlidx := bytes.IndexByte(b, '\n')
+	if cridx > 0 {
+		idx = cridx
+	} else {
+		cridx = len(b) + 1
+	}
+	if nlidx > 0 && nlidx < cridx {
+		idx = nlidx
+	}
+	if idx <= 0 {
+		return string(b), true
+	}
+	return string(b[:idx]), false
+}
